# Faster whitespace detection