@davecheney 2020-06-28
It’s not actually faster :(